backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
sqladmin>=0.16.0
asyncpg>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
email-validator>=2.1.0
jinja2>=3.1.3
python-multipart>=0.0.9
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
itsdangerous>=2.1.2
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
alembic>=1.13.0
